Fix drawer width bug
If a double or float was sent instead of an int, the default of -1 would
be used. Resulting in the default drawer width.

+++ android/app/src/main/java/com/reactnativenavigation/params/parsers/SideMenuParamsParser.java
@@ -23,7 +23,7 @@ class SideMenuParamsParser extends Parser {
         result.screenId = sideMenu.getString("screenId");
         result.navigationParams = new NavigationParams(sideMenu.getBundle("navigationParams"));
         result.disableOpenGesture = sideMenu.getBoolean("disableOpenGesture", false);
-        result.drawerWidth = sideMenu.getInt("drawerWidth", -1);
+        result.drawerWidth = (int)sideMenu.getDouble("drawerWidth", -1);
         result.side = side;
         return result;
     }
